# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Playita de Don Concho - Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    Role:
      type: string
      enum: [ADMIN, WAITER, CASHIER]
...
    User:
      type: object
      required: [id, name, role, active]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
    CreateUserRequest:
      type: object
      required: [name, password, role]
      properties:
        name: { type: string }
        password: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        password: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
...